refactor: 파일명 및 폴더 구조 개선
