+++ src/featured/members/actions/members.action.ts
@@ -1,0 +1,1 @@
+'use server'
